Correct test - using MongoTemplate + new ShelConvertUtil.class for Shell Convert

+++ src/main/java/learn/library/service/LibraryImpl.java
@@ -55,7 +55,11 @@ public class LibraryImpl implements Library {
     @Override
     @Transactional
     public void deleteBookById(String id) {
-        bookRepository.deleteById(id);
+        Optional<Book> optionalBook = bookRepository.findById(id);
+        if (optionalBook.isPresent()) {
+            bookRepository.deleteById(id);
+            commentRepository.deleteAllByBook(optionalBook.get());
+        }
     }
 
     @Override
